Replace Wait with Awaitility and CleanableDockerComposeRule new JUnit5 Docker-Compose-Rue extension. Migrate to JUnit5. Merga all IT test classes in one. Drop the Influx/Postgres/Rabbit plan
