+++ pom.xml
@@ -96,7 +96,7 @@
 		<apache-directory-server.version>1.5.5</apache-directory-server.version>
 		<codearte-props2yml.version>0.5</codearte-props2yml.version>
 		<json-unit.version>1.25.0</json-unit.version>
-		<docker-compose-rule-junit4.version>1.4.2</docker-compose-rule-junit4.version>
+		<docker-compose-rule.version>1.4.2</docker-compose-rule.version>
 	</properties>
 	<modules>
 		<module>spring-cloud-dataflow-configuration-metadata</module>
@@ -257,8 +257,8 @@
 			</dependency>
 			<dependency>
 				<groupId>com.palantir.docker.compose</groupId>
-				<artifactId>docker-compose-rule-junit4</artifactId>
-				<version>${docker-compose-rule-junit4.version}</version>
+				<artifactId>docker-compose-junit-jupiter</artifactId>
+				<version>${docker-compose-rule.version}</version>
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
